flask/requirements.txt
gunicorn==19.9.0
gevent==1.4.0
flask==1.0.2
flask-sqlalchemy
flask-mysqldb
flask-cors
pymysql
pycryptodome
requests
pillow